fix(plugin): restore the two required lanes broken by #2200
#2200 merged with 'Fast (Linux x86_64)' and 'Rust quality' both already
failing, leaving main red for every PR in the repo. Bisected: the ABI
tests pass at 4b60b9903 (its parent) and fail at 71959075a.

Two of the three breakages are unambiguous and fixed here.

1. clippy, -D warnings: `raw_device_type as u32` casts u32 to u32.

2. Compute now resolves each input's real allocator residency via
   GetTensorMemoryInfo / MemoryInfoGetDeviceType / MemoryInfoGetName /
   MemoryInfoGetId and fails closed when any is null. The plugin ABI
   test's mock OrtApi -- documented as providing 'all the functions our
   Compute path needs' -- was never given them, so all 7 Compute tests
   returned an error status. Added; the mocks report a CPU allocator so
   device_from_memory_info resolves DeviceType::Cpu as before.

The third is NOT fixed here because it is a design decision belonging to
#2200's author: four new cuda-plugin fixtures declare opset 13/17
against onnx-std's fixture floor of 24. Those opsets are deliberate --
ReduceSum axes-as-input from 13, DFT default axis 1 at 17 -- so raising
them would destroy what the fixtures test. The guard's own remedy is to
build such graphs in-memory in the test that needs them, which is a
refactor of CUDA-gated tests that cannot be validated here.

+++ crates/onnx-runtime-ep-plugin/src/kernel_ctx.rs
@@ -234,7 +234,7 @@ pub(crate) unsafe fn device_from_memory_info(
         "qnn" => DeviceType::Qnn,
         "openvino" => DeviceType::OpenVino,
         _ if raw_device_type == ort::OrtMemoryInfoDeviceType_CPU => DeviceType::Cpu,
-        _ => DeviceType::Custom(raw_device_type as u32),
+        _ => DeviceType::Custom(raw_device_type),
     };
     Ok(DeviceId::new(device_type, index))
 }
